Redesign kernel structure and clean code
Changes to core design were inevitable to make things work charming
under file system. It was a quite heavy job for me since I didn't have
enough knowledge to connect the whole things under file system. Revising
file system is still going on.

Add stack sentinel
  It checks stack overflow of `current` every scheduling, specifically
  in `runqueue_add()`, if `CONFIG_DEBUG` set.

+++ mach/stm32/include/io.h
@@ -41,7 +41,7 @@
 #define FLASH_WRITE_WORD(addr, data)	{ \
 		*(volatile unsigned short int *)addr = (unsigned short int)data; \
 		while (FLASH_SR & 1); /* Check BSY bit, need timeout */ \
-		*(volatile unsigned short int *)(addr+2) = (unsigned short int)(data >> 16); \
+		*(volatile unsigned short int *)((unsigned int)addr+2) = (unsigned short int)(data >> 16); \
 		while (FLASH_SR & 1); /* Check BSY bit, need timeout */ \
 	}
 #define FLASH_LOCK()	(FLASH_CR |= 0x80)
